Vibe refactor
