+++ deep_quoridor/src/agents/__init__.py
@@ -29,5 +29,9 @@ class Agent:
         raise NotImplementedError("You must implement the get_action method")
 
 
+__all__ = ["RandomAgent", "SimpleAgent", "Agent", "DQNAgent"]
+
 from agents.random import RandomAgent  # noqa: E402, F401
 from agents.simple import SimpleAgent  # noqa: E402, F401
+from agents.agent import Agent
+from agents.flat_dqn import DQNAgent
